prospecting-agent/requirements.txt
googlemaps==4.10.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
openpyxl==3.1.2
openai==1.30.1
python-dotenv==1.0.1
tqdm==4.66.4
